Use - instead of __fake_global
It's not actually a valid function name in C, but cscope doesn't barf on it.
It displays much more cleanly, and isn't searchable anyways since cscope's UI
doesn't let you enter it and vim doesn't recognize it as a word character. It
conflicts with people overriding subtraction on their ruby classes, but again
those aren't searchable anyways.