Refactor and remove StarScope::Record
Most of it was export-related, which belongs in the new StarScope::Export
module. There was also some export-related code left in the DB (specifically the
CSCOPE_GLOBAL_HACK bits) so move that over. This means that *all* of the
export-related code is *finally* in one place!

The remainder of Record didn't deserve its own module, so just mix it in with
the now-much-more-managable DB.