cscope: fix inline function definitions
Apparently cscope does not like to see the end-of-function mark (`\t}`) in the
middle of a line without any preceding plain text (i.e. when the preceding item
on the line is also a symbol). This can actually happen in ruby because of the
magic 'global' function hack I added (working around a different cscope
limitation). As a workaround, add dummy whitespace if this happens.

Goddamn I wish vim supported a less insane file format for this stuff.