Ensure that `cscope_plaintext` never returns ''
If it does, we'll accidentally create a cscope file with two consecutive
newlines in the middle of a single 'real' source line. This is effectively a
corrupt file that only ever shows up if you search for a symbol on that line,
because cscope's parsing logic is so weird.

I had to read cscope's source to track this one down. The annoying case was in a
particular javascript file, but this probably solves a whole raft of heisenbugs
in ruby and go as well.